feat: add collections functionality
Allow user to create, edit and delete their own collections and to add or remove recipes. Create a collection details page to view dish cards associated with that collection.

+++ WebApplication1/Controllers/CollectionsController.cs
@@ -2,9 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using MyRecipes.Areas.Identity.Pages.Account;
 using MyRecipes.Data;
 using MyRecipes.Models;
 
@@ -13,22 +15,29 @@ namespace MyRecipes.Controllers
     public class CollectionsController : Controller
     {
         private readonly AppDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public CollectionsController(AppDbContext context)
+        public CollectionsController(AppDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         // GET: Collections
         public async Task<IActionResult> Index()
         {
-            var appDbContext = _context.Collection.Include(c => c.User);
-            return View(await appDbContext.ToListAsync());
+            var currentUserId = _userManager.GetUserId(HttpContext.User);
+
+            var collections = await _context.Collection.Include(c => c.DishCollections).Where(c => c.UserId == currentUserId).ToListAsync();
+
+            return View(collections);
         }
 
         // GET: Collections/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            var currentUserId = _userManager.GetUserId(HttpContext.User);
+
             if (id == null)
             {
                 return NotFound();
@@ -43,14 +52,16 @@ namespace MyRecipes.Controllers
             {
                 return NotFound();
             }
-
+            if (collection.UserId != currentUserId)
+            {
+                return RedirectToAction(nameof(Index));
+            }
             return View(collection);
         }
 
         // GET: Collections/Create
         public IActionResult Create()
         {
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id");
             return View();
         }
 
@@ -59,32 +70,38 @@ namespace MyRecipes.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,CollectionName,UserId")] Collection collection)
+        public async Task<IActionResult> Create([Bind("Id,CollectionName")] Collection collection)
         {
+            var currentUserId = _userManager.GetUserId(HttpContext.User);
+            collection.UserId = currentUserId;
             if (ModelState.IsValid)
             {
                 _context.Add(collection);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", collection.UserId);
             return View(collection);
         }
 
         // GET: Collections/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var collection = await _context.Collection.FindAsync(id);
+            var currentUserId = _userManager.GetUserId(HttpContext.User);
+            if(currentUserId != collection.UserId)
+            {
+                return RedirectToAction(nameof(Index));
+            }
             if (collection == null)
             {
                 return NotFound();
             }
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", collection.UserId);
             return View(collection);
         }
 
@@ -95,12 +112,14 @@ namespace MyRecipes.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,CollectionName,UserId")] Collection collection)
         {
+            var currentUserId = _userManager.GetUserId(HttpContext.User);
+
             if (id != collection.Id)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && collection.UserId == currentUserId)
             {
                 try
                 {
@@ -127,13 +146,13 @@ namespace MyRecipes.Controllers
         // GET: Collections/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var collection = await _context.Collection
-                .Include(c => c.User)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (collection == null)
             {
@@ -148,8 +167,10 @@ namespace MyRecipes.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var currentUserId = _userManager.GetUserId(HttpContext.User);
+
             var collection = await _context.Collection.FindAsync(id);
-            if (collection != null)
+            if (collection != null && collection.UserId == currentUserId)
             {
                 _context.Collection.Remove(collection);
             }
